fixed parallel runner to check for #686
